mv torch_object_detection object_detection (and clean up the training notebook)
