+++ object_detection/ycb_labels.yaml
@@ -1,0 +1,12 @@
+1: 001_chips_can
+2: 025_mug
+3: 026_sponge
+4: 004_sugar_box
+5: 013_apple
+6: 029_plate
+7: 031_spoon
+8: 003_cracker_box
+9: 011_banana
+10: 030_fork
+11: 024_bowl
+12: 037_scissors
