--- torch_object_detection/ycb_labels.yaml
@@ -1,5 +1,0 @@
-category_names: ['__background__', 'can', 'sponge']
-label_map:
-    '__background__': 0
-    'can': 1
-    'sponge': 2
